+++ .storybook/main.ts
@@ -1,3 +1,4 @@
+import { dirname, join } from "path";
 import MiniCssExtractPlugin from "mini-css-extract-plugin";
 import type { StorybookConfig } from "@storybook/react-webpack5";
 import { VanillaExtractPlugin } from "@vanilla-extract/webpack-plugin";
@@ -12,14 +13,14 @@ const config: StorybookConfig = {
     },
   },
   addons: [
-    "@storybook/addon-links",
-    "@storybook/addon-essentials",
-    "@storybook/addon-interactions",
-    "storybook-dark-mode",
-    "@storybook/addon-mdx-gfm",
+    getAbsolutePath("@storybook/addon-links"),
+    getAbsolutePath("@storybook/addon-essentials"),
+    getAbsolutePath("@storybook/addon-interactions"),
+    getAbsolutePath("storybook-dark-mode"),
+    getAbsolutePath("@storybook/addon-mdx-gfm"),
   ],
   framework: {
-    name: "@storybook/react-webpack5",
+    name: getAbsolutePath("@storybook/react-webpack5"),
     options: {},
   },
   staticDirs: [
@@ -57,3 +58,7 @@ const config: StorybookConfig = {
   },
 };
 export default config;
+
+function getAbsolutePath(value: string): any {
+  return dirname(require.resolve(join(value, "package.json")));
+}